device-agent/pyproject.toml
[tool.poetry]
name = "device-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.6.1"
tinkerforge = "^2.1.30"
certifi = "^2022.12.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
